refactor(chat): split thread action helpers
